+++ src/main/java/com/uniovi/services/UsersService.java
@@ -18,7 +18,7 @@ import com.uniovi.repositories.UsersRepository;
 public class UsersService {
 	@Autowired
 	private UsersRepository usersRepository;
-	
+
 	@Autowired
 	private BCryptPasswordEncoder bCryptPasswordEncoder;
 
@@ -26,28 +26,28 @@ public class UsersService {
 	public void init() {
 	}
 
-	public List<User> getUsers() {
+	public List<User> getUsers(User user) {
 		List<User> users = new ArrayList<User>();
 		usersRepository.findAll().forEach(users::add);
+		users.remove(users.indexOf(user));
 		return users;
 	}
-	
+
 	public User getUserByEmail(String email) {
 		return usersRepository.findByEmail(email);
-		}
-
+	}
 
 	public User getUser(Long id) {
 		return usersRepository.findById(id).get();
 	}
-	
+
 	public void addUser(User user) {
 		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
 		usersRepository.save(user);
 	}
 
-
-	public void deleteUser(Long id) {
-		usersRepository.deleteById(id);
+	public void deleteUser(Long id, User actual) {
+		if (id != actual.getId())
+			usersRepository.deleteById(id);
 	}
 }
